user validations working except pregnant
